requirements.txt
fastapi
uvicorn
pandas
numpy
pydantic
pydantic-settings
langchain==0.1.20
langchain-community==0.0.38
langchain-groq
sentence-transformers
chromadb
python-dotenv
scikit-learn
sqlalchemy
psycopg2-binary
